fix!(ListaDeseos-Usuario): Corregir tipo de dato llave

+++ backend/src/main/java/com/debede/backend/repository/ListaDeseos_UsuariosRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ListaDeseos_UsuariosRepository extends JpaRepository<ListaDeseos_Usuarios, Long> {
+public interface ListaDeseos_UsuariosRepository extends JpaRepository<ListaDeseos_Usuarios, Integer> {
 
 }
